+++ packages/core/src/transaction/Clause.ts
@@ -1,4 +1,3 @@
-import { InvalidDataType } from '@vechain/sdk-errors';
 import { ERC721_ABI, VIP180_ABI } from '../utils';
 import {
     abi,
@@ -12,8 +11,10 @@ import {
 } from '../vcdm';
 import { Hex } from '../vcdm/Hex';
 import { HexInt } from '../vcdm/HexInt';
+import { InvalidDataType } from '@vechain/sdk-errors';
 import type { DeployParams } from './DeployParams';
-import { type ClauseOptions, type TransactionClause } from './types';
+import type { TransactionClause } from './TransactionClause';
+import type { ClauseOptions } from './ClauseOptions';
 
 /**
  * This class represent a transaction clause.
