+++ packages/core/src/transaction/index.ts
@@ -1,8 +1,8 @@
 export * from './Clause';
+export * from './Transaction';
+export * from './transaction-handler';
+export * from '../utils/const/network';
 export type * from './ClauseOptions';
 export type * from './DeployParams';
 export type * from './TransactionBody';
 export type * from './TransactionClause';
-export * from './Transaction';
-export * from './transaction-handler';
-export * from '../utils/const/network';
